refactoring amplifier
